Now using aarch64-darwin fakedir-universal
Pretty much all derivations used and cached by Nixie are meant to not
have any Nix store dependencies, and due to the breadth of architectures
that need to come together to synthesize our top-level derivations,
standardizing around the aarch64-darwin version of the derivation just
makes more sense -- consistent hashes for Intel Mac users, local build
ability for AArch64 users.

Funnily enough, the root derivation for static binaries will be
x86_64-linux, should it only be because it's kinda "the default" in the
Nix ecosystem.