Now building static Nix from upstream flake
This allows us to leverage a static binary build that actually works,
including on Darwin, and comes with the added benefit of a much
simplified Nix version model -- whatever's written on top of the
flake.nix, that's the version you'll be shipping.

Only time will tell if this approach is more robust, and ideally the
patch set will be cleaned up and upstreamed.